requirements.txt
pyyaml
pandas
scikit-image
matplotlib
torchvision
pytorch
numpy
h5py


